+++ js/utilities.js
@@ -1,0 +1,25 @@
+export {
+  getRandomInteger,
+  getRandomFloat
+};
+
+function getRandomInteger(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  if (max <= min) {
+    return null;
+  } else if (max < 0 || min < 0) {
+    return null;
+  }
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+function getRandomFloat(min, max, fraction) {
+  if (max <= min) {
+    return null;
+  } else if (max < 0 || min < 0) {
+    return null;
+  }
+  const randomNumber = (Math.random() * (max - min)) + min;
+  return +randomNumber.toFixed(fraction);
+}
